+++ lib_ts/index.js
@@ -42,8 +42,10 @@ var WizardNav_1 = require("./components/WizardNav");
 exports.WizardNav = WizardNav_1.default;
 var WizardStep_1 = require("./components/WizardNav/WizardStep");
 exports.WizardStep = WizardStep_1.default;
-var Datagrid_1 = require("./components/Datagrid");
-exports.Datagrid = Datagrid_1.default;
+var DataTable_1 = require("./components/Datagrid/DataTable");
+exports.DataTable = DataTable_1.default;
+var DataTable_2 = require("./components/Datagrid/DataTable"); // BC compatibility
+exports.Datagrid = DataTable_2.default;
 var DataRow_1 = require("./components/Datagrid/DataRow");
 exports.DataRow = DataRow_1.default;
 var DataCell_1 = require("./components/Datagrid/DataCell");
