+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "mergado-ui-kit",
-  "version": "1.5.6-dev",
+  "version": "1.5.7-dev",
   "description": "",
   "main": "dist/bundle.js",
   "scripts": {
