+++ src/components/Datagrid/index.tsx
@@ -1,31 +1,4 @@
-import * as React from "react"
-import {prefix} from "config"
-
-export interface Props {
-    style?: any
-    addClass?: string
-}
-export interface State {
-}
-
-class DataTable extends React.Component<Props, State> {
-
-    readonly name = prefix + "datagrid";
-
-    public static defaultProps: Props = {
-        style: {},
-        addClass: ''
-    }
-
-    render() {
-        const { addClass, style } = this.props
-        let className = `${this.name} ${this.props.addClass}`
-        return (
-            <table className={className} style={style}>
-                {this.props.children}
-            </table>
-        )
-    }
-}
-
-export default DataTable
+export {default as DataTable} from "components/Datagrid/DataTable"
+export {default as DataRow} from "components/Datagrid/DataRow"
+export {default as DataCell} from "components/Datagrid/DataCell"
+export {default as DataHeader} from "components/Datagrid/DataHeader"
