+++ src/components/Forms/index.tsx
@@ -1,0 +1,10 @@
+export {default as Autocomplete} from "components/Forms/Autocomplete"
+export {default as Button} from "components/Forms/Button"
+export {default as Checkbox} from "components/Forms/Checkbox"
+export {default as CheckboxContainer} from "components/Forms/CheckboxContainer"
+export {default as ColorPicker} from "components/Forms/ColorPicker"
+export {default as Radio} from "components/Forms/Radio"
+export {default as Range} from "components/Forms/Range"
+export {default as Select} from "components/Forms/Select"
+export {default as Textarea} from "components/Forms/Textarea"
+export {default as TextInput} from "components/Forms/TextInput"
