+++ src/index.ts
@@ -23,7 +23,8 @@ export {default as NavLink} from "./components/TopNav/NavLink"
 export {default as WizardNav} from "./components/WizardNav"
 export {default as WizardStep} from "./components/WizardNav/WizardStep"
 
-export {default as Datagrid} from "./components/Datagrid"
+export {default as DataTable} from "./components/Datagrid/DataTable"
+export {default as Datagrid} from "./components/Datagrid/DataTable" // BC compatibility
 export {default as DataRow} from "./components/Datagrid/DataRow"
 export {default as DataCell} from "./components/Datagrid/DataCell"
 export {default as DataHeader} from "./components/Datagrid/DataHeader"
